fix store persist
